\name{RKindle-package}
\alias{RKindle-package}
\alias{RKindle}
\docType{package}
\title{Kindle collection management}
\description{Creating JSON collection files for Kindle.}
\details{
\tabular{ll}{
Package: \tab RKindle\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2013-09-28\cr
License: \tab GPL >= 2\cr
}
}
\author{Juraj Medzihorsky}
\references{}
\keyword{ package }
\seealso{
	\code{\link[SoDA:SoDA-package]{SoDA}}
	\code{\link[rjson:rjson-package]{rjson}}
}
\examples{}
